oxr v0.4.0 OXR

A thin wrapper for the Open Exchange Rates API. Requires an app id, which you can receive at https://openexchangerates.org

Types

converted()

converted() :: number

currencies()

currencies() :: %{optional(String.t) => String.t}

error_description()

error_description() :: String.t

error_message()

error_message() :: String.t

http_status()

http_status() :: 200..511

rate()

rate() :: number

rates()

rates() :: %{optional(String.t) => rate}

timestamp()

timestamp() :: pos_integer

Functions

convert(app_id, value, from, to)

convert(String.t, number, String.t, String.t) ::
  {:ok, timestamp, rate, converted} |
  {:error, http_status, error_message, error_description} |
  {:error, error_description}

Convert money value from one currency to another

currencies(show_experimental \\ false)

currencies(boolean) ::
  {:ok, currencies} |
  {:error, http_status, error_message, error_description} |
  {:error, error_description}

Get all currencies available

historical(app_id, date, base \\ "", symbols \\ [])

Get historical exchange rates for any date available Default base is USD

latest(app_id, base \\ "", symbols \\ [])

latest(String.t, String.t, [String.t, ...]) ::
  {:ok, timestamp, rates} |
  {:error, http_status, error_message, error_description} |
  {:error, error_description}

Get latest exchange rates Default base is USD

time_series(app_id, start, stop, base \\ "", symbols \\ [])

time_series(String.t, String.t, [String.t, ...], String.t, String.t) ::
  {:ok, rates} |
  {:error, http_status, error_message, error_description} |
  {:error, error_description}

Get historical exchange rates for a given time period Default base is USD

usage(app_id)

usage(String.t) ::
  {:ok, %Data{app_id: term, plan: term, status: term, usage: term}} |
  {:error, http_status, String.t, String.t} |
  {:error, String.t}

Get basic plan information and usage statistics Warning: this endpoint is still in beta